moving FlashingText to widgets.py
